+++ src/lib/helpers/uploader.helper.ts
@@ -1,36 +1,55 @@
 import imageCompression from 'browser-image-compression';
 import { keccak256 } from 'ethers/lib/utils';
+import imageSize from 'image-size';
 import { AddResult } from 'ipfs-core-types/src/root';
 import { ImportCandidate } from 'ipfs-core-types/src/utils';
 import { create, Options } from 'ipfs-http-client';
+import Jimp from 'jimp';
 
 import { LSP3ProfileImage } from '../interfaces';
+import { ImageBuffer } from '../interfaces/lsp3-profile';
 import { ProfileUploadOptions } from '../interfaces/profile-upload-options';
 
 export const sizes = [1800, 1024, 640, 320, 180];
 export async function imageUpload(
-  givenFile: File,
+  givenFile: File | ImageBuffer,
   uploadOptions: ProfileUploadOptions
 ): Promise<LSP3ProfileImage[]> {
-  const isImage = givenFile.type?.substr(0, 6) === 'image/';
+  const type = 'type' in givenFile ? givenFile.type : givenFile.mimeType;
+  const isImage = type?.substr(0, 6) === 'image/';
   if (!isImage) {
-    throw new Error(`File provided is of type "${givenFile.type}".`);
+    throw new Error(`File type: '${type}' does not start with 'image/'`);
   }
 
   return Promise.all(
     sizes.map(async (size) => {
-      const img = await imageCompression(givenFile, {
-        maxWidthOrHeight: size,
-        useWebWorker: true,
-      });
+      let imgToUpload, imgBuffer, width: number, height: number;
 
-      const imgBuffer = new Uint8Array(await img.arrayBuffer());
-      const loadedImg = await imageCompression.drawFileInCanvas(img);
-      const uploadResponse = await ipfsUpload(img, uploadOptions.ipfsClientOptions);
+      if ('buffer' in givenFile) {
+        imgBuffer = await resizeBuffer(givenFile.buffer, givenFile.mimeType, size);
+        imgToUpload = imgBuffer;
+
+        const resizedDimensions = imageSize(imgBuffer);
+        height = resizedDimensions.height;
+        width = resizedDimensions.width;
+      } else {
+        imgToUpload = await imageCompression(givenFile, {
+          maxWidthOrHeight: size,
+          useWebWorker: true,
+        });
+
+        imgBuffer = new Uint8Array(await imgToUpload.arrayBuffer());
+        const loadedImg = await imageCompression.drawFileInCanvas(imgToUpload);
+
+        height = loadedImg[0].height;
+        width = loadedImg[0].width;
+      }
+
+      const uploadResponse = await ipfsUpload(imgToUpload, uploadOptions.ipfsClientOptions);
 
       return {
-        width: loadedImg[0].width,
-        height: loadedImg[0].height,
+        width: height,
+        height: width,
         hashFunction: 'keccak256(bytes)',
         hash: keccak256(imgBuffer),
         url: 'ipfs://' + uploadResponse.cid.toString(),
@@ -48,7 +67,7 @@ export async function ipfsUpload(file: ImportCandidate, options: Options): Promi
 
 export async function prepareImageForLSP3(
   uploadOptions?: ProfileUploadOptions,
-  image?: File | LSP3ProfileImage[]
+  image?: File | ImageBuffer | LSP3ProfileImage[]
 ): Promise<LSP3ProfileImage[]> | null {
   let lsp3Image: LSP3ProfileImage[] | null;
 
@@ -59,4 +78,9 @@ export async function prepareImageForLSP3(
   }
 
   return lsp3Image;
+}
+
+export async function resizeBuffer(buffer: Buffer, format: string, size: number): Promise<Buffer> {
+  const image = await Jimp.read(buffer);
+  return image.scaleToFit(size, size).getBufferAsync(format);
 }