+++ src/lib/interfaces/lsp3-profile.ts
@@ -24,9 +24,22 @@ export interface LSP3ProfileImage {
   url: string;
 }
 
+/**
+ * @example
+ *```javascript
+ *{
+ *  name: "My Universal Profile",
+ *  description: "My cool Universal Profile",
+ *  profileImage: myLocalFile,
+ *  backgroundImage: myLocalFile,
+ *  tags: ['Fashion', 'Design'],
+ *  links: [{ title: "My Website", url: "www.my-website.com" }],
+ *};
+ *```
+ */
 export interface ProfileDataBeforeUpload {
-  profileImage?: File | LSP3ProfileImage[];
-  backgroundImage?: File | LSP3ProfileImage[];
+  profileImage?: File | ImageBuffer | LSP3ProfileImage[];
+  backgroundImage?: File | ImageBuffer | LSP3ProfileImage[];
   name: string;
   description: string;
   links?: LSP3ProfileLink[];
@@ -36,4 +49,16 @@ export interface ProfileDataBeforeUpload {
 export interface LSP3ProfileDataForEncoding {
   profile: LSP3ProfileJSON;
   url: string;
+}
+
+export interface ImageBuffer {
+  buffer: Buffer;
+  mimeType: SupportedImageBufferFormats;
+}
+
+export enum SupportedImageBufferFormats {
+  MIME_PNG = 'image/png',
+  MIME_BMP = 'image/bmp',
+  MIME_JPEG = 'image/jpeg',
+  MIME_GIF = 'image/gif',
 }